pyproject.toml
[project]
name = "thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets==2.20.0",
    "ipycytoscape>=1.3.3",
    "library>=3.0.49",
    "loguru>=0.7.3",
    "matplotlib>=3.10.0",
    "neo4j>=5.27.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "torch>=2.5.1",
    "yfiles-jupyter-graphs-for-neo4j>=1.6.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
